CLEANUP | Move suggestion generation to own component
